pyproject.toml
[tool.poetry]
name = "tmdbwrapper"
version = "0.1.0"
description = ""
authors = ["Purna Nugraha <ws.purnasatria@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
vcrpy = "^6.0.1"

[tool.pyright]
venvPath = "."
venv = ".venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
